Fixed reading project build properties for packages.config based project
Currently we had 2 different implementation to read project build properties where one reads from dte project and another reads from IVsBuildPropertyStorage. Except for classic PackageReference, we should always read build properties from dte project since most of these properties will only available in dte project instance. But with our LSL changes, we regressed this behavior and started reading build properties from IVsBuildPropertyStorage.

This PR fixes this regression by merging these two different implementations into one class where IVsBuildPropertyStorage will fallback to dte project. It will not impact classic or netcore PackageReference, and will just fix packages.config based projects.

Fixes #5880
