+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/NuGet.PackageManagement.VisualStudio.csproj
@@ -211,8 +211,7 @@
     <Compile Include="ProjectServices\DeferredProjectServicesProxy.cs" />
     <Compile Include="Projects\IVsProjectAdapterProvider.cs" />
     <Compile Include="Services\IVsProjectThreadingService.cs" />
-    <Compile Include="ProjectServices\VsCoreProjectBuildProperties.cs" />
-    <Compile Include="ProjectServices\VsManagedLanguagesProjectBuildProperties.cs" />
+    <Compile Include="ProjectServices\VsProjectBuildProperties.cs" />
     <Compile Include="ProjectServices\VsManagedLanguagesProjectSystemServices.cs" Condition="'$(VisualStudioVersion)' == '15.0'" />
     <Compile Include="Projects\VsProjectAdapter.cs" />
     <Compile Include="Projects\VsProjectAdapterProvider.cs" />
